minor type-change
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@934 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/examples/elliptic/ellipticmodel.hh
@@ -139,7 +139,7 @@ public:
         }
 
   //! the coefficient for robin boundary condition
-  inline int alpha()
+  inline double alpha()
         { 
           return 1.0;
         }
@@ -362,7 +362,7 @@ public:
         }
 
   //! the coefficient for robin boundary condition
-  inline int alpha()
+  inline double alpha()
         { 
           return 1.0;
         }
@@ -567,7 +567,7 @@ public:
         }
 
   //! the coefficient for robin boundary condition
-  inline int alpha()
+  inline double alpha()
         { 
           return 1.0;
         }
